finish new part

+++ src/IR/constant.java
@@ -5,7 +5,7 @@ public class constant extends entity{
         STRING,INT,VOID,BOOL
     }
     public Genre genre;
-    private int intValue;
+    private long intValue;
     private boolean boolValue;
     private String stringValue;
 
@@ -19,7 +19,7 @@ public class constant extends entity{
         this.genre = Genre.BOOL;
         this.boolValue = flag;
     }
-    public constant(int value) {
+    public constant(long value) {
         super();
         this.intValue = value;
         genre = Genre.INT;
@@ -31,7 +31,7 @@ public class constant extends entity{
         genre = Genre.STRING;
     }
 
-    public int getIntValue() { return intValue;}
+    public long getIntValue() { return intValue;}
     public String getStringValue() { return stringValue;}
     public boolean getBoolValue() {return boolValue;}
 }
